Optimized multi-byte serial writes and peek
Transitioned to using a buffered write model, rather than blocking the
caller.  Also optimized the peek() function so that interrupts no longer
needed to be masked during its invocation.  Refactored code to reset
UART if framing or other errors are encountered; this case appears to
occur infrequently when communicating with the RoboRIO.