Fix:  STM32 USB ifx Stuck transmitting on RoboRIO Reset
Since the VBUS signal from the USB connector is not connected to the
STM32 chip in this design, it's not able to detect VBUS removed/added
events.  This leads to a software error (sometimes) when the host device
is reset.  This modification detects a reset (when the host device comes
back online) and re-initializes the USB circuitry in this case.  This
fixes the issue experienced periodically upon RoboRIO resets, and also
vastly improves the issue on the PC when the USB cable is
removed/reinserted during operation.