Lengthened Factory Calibration Period
Previously, sometimes the factory calibration would terminate before the
MPU had settled on the best accelerometer bias value (previously it
waited for the first change in calibration biases by the DMP).  This
period of time was lengthed, by increasing the number of changes in cal
bias that are looked for.