Increase USB Reset timeout
Increased the USB reset timeout from 1 to 2 seconds.   This change was
necessary as during testing of the (faster) C++ libraries, sometimes the
RoboRio would assign a new device name to the navX USB device with a
shorter wait during the USB reset cycle.  Now that this is increased to
2 second (this time period is applied on navX startup), each of the 3
libraries (and the navX MXP UI on the PC) all recover gracefully from
device resets.