Added support for I2C transfers w/Masters using repeated start conditions
The previous implementation required the master reader to issue a
two-byte write (start address, len) then a read transaction.  In order
to work with i2c masters requiring support for a "repeated start" read
where the length is not specified, support for the one-byte read request
followed by the start condition was added.  In addition, support for i2c
masters which do not terminate the read with a stop is also now
supported.