Fixed loop and linear rotate-to-angle and drive-straight examples
- Affected loop op modes had a stop() method added which closed the AHRS
device which correctly deregistered callbacks and allowed the PID
controller to be garbage collected.
- Affected linear op modes were reworked to ensure that the
InterruptedException() and interruption condition were correctly
handled, and added code to close the AHRS device as in the loop op mode
samples.  As a result, the robot controller isn't exited any more due to
the linear op mode code taking too long to quit.