Modifications to fix issues with AHRS.getAngle()
Rewrote the ContinuousAngleTracker, using portions of the code provided
by Joe Ross, FRC Team 330.  Also added support for integration of yaw
reset commands into the tracker, as well as support for an "adjustment
angle" to the angle returned from getAngle().

Also includes a bugfix to the length in the decode function of the
AHRSPosTSUpdate method (C++ only).