Update to help with USB Lock-up issue
I changes the zero notifier:
I added code to make the action engine look for a change to fire the
notifier to the main loop.
I added code to make it so the operator can only fire a zero notifier at
1 second intervals.
I changed the way the loops are metered to get to the desitred loop
rate. It is more stable and less complicated.
I changed the way the serial and USB loops are reading data to make it
more effiecent.
I wait to interpert the data until I get a recognized pact from the
device. Then I read all of the data and process it.