Updated RoboRIO C++ Libraries to use software-based yaw resets
For most FRC software usage, software-based yaw resets will provided
quicker response and avoid some of the delays inherent with Board-level
yaw resets, and also avoid some of the issues with USB-based
communication on the roborio, which is intolerant in certain cases of
USB writes while USB data is being received.

This includes enhancements to ensure that board-level resets are
detected when serial/USB communication is used.